pyproject.toml
[project]
name = "snowberry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cfbd>=4.5.2",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
]
